# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yappr Kokoro v1 TTS + Whisper STT Server",
    "version": "0.1.0"
  },
  "paths": {
    "/voices": {
      "get": {
        "summary": "Get Voices",
        "description": "List American English Kokoro v1 voice ids.\n\nReturns ``application/json`` with a ``voices`` array of strings (``af_*``, ``am_*``)\nfor **hexgrad/Kokoro-82M** when using ``lang_code=a`` in the pipeline.",
        "operationId": "get_voices_voices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/synthesize": {
      "post": {
        "summary": "Synthesize",
        "description": "Synthesize speech from text using the loaded Kokoro pipeline.\n\n**Response:** ``audio/wav`` bytes (HTTP 200). Returns 503 if the TTS model failed to load.",
        "operationId": "synthesize_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transcribe": {
      "post": {
        "summary": "Transcribe",
        "description": "Transcribe uploaded audio to text (and language metadata).\n\n**Response:** JSON with ``text``, ``language``, and ``probability``. Returns 503 if STT is unavailable.",
        "operationId": "transcribe_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_transcribe_transcribe_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Audio upload (e.g. WAV). Transcribed with faster-whisper ``base.en`` when the STT model loaded at startup."
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_transcribe_transcribe_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SynthesizeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Plain text to speak."
          },
          "voice": {
            "type": "string",
            "title": "Voice",
            "description": "Kokoro v1 voice id (American English: ``af_*`` / ``am_*``). See ``GET /voices`` for the supported list.",
            "default": "af_bella"
          },
          "speed": {
            "type": "number",
            "title": "Speed",
            "description": "Speaking-rate multiplier (1.0 = model default).",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SynthesizeRequest",
        "description": "JSON body for ``POST /synthesize``.\n\nProduces linear PCM in a WAV container via Kokoro 82M (``hexgrad/Kokoro-82M``, ``lang_code=a``)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}